requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
bandit==1.7.5
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
flake8==6.0.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
invoke==2.1.1
isort==5.12.0
markdown-it-py==2.2.0
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.4
mypy==1.3.0
mypy-extensions==1.0.0
numpy==1.24.3
openai==0.27.6
packaging==23.1
pathspec==0.11.1
pbr==5.11.1
platformdirs==3.5.1
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.15.1
python-dotenv==1.0.0
PyYAML==6.0
requests==2.30.0
rich==13.3.5
semantic-kernel==0.2.7.dev0
smmap==5.0.0
stevedore==5.0.0
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==2.0.2
uuid6==2023.5.2
yarl==1.9.2
